+++ voxygen/src/render/pipelines/particle.rs
@@ -120,6 +120,7 @@ pub enum ParticleMode {
     Ice = 21,
     LifestealBeam = 22,
     CultistFlame = 23,
+    StaticSmoke = 24,
 }
 
 impl ParticleMode {