+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generated a net world map
 - Overhauled clouds for more verticality and performance
 - New tooltip for items with stats comparison
+- Improved bow feedback, added arrow particles
 
 ### Removed